cloudbuild: limit image release tags to semver only
When doing cloudbuilds for tagged releases, limit images to only a
single tag that represents the semver tag.

The registry.k8s.io tooling uses scripts to coordinate which images and
tags are promoted from staging to release. The script don't handle image
digests with multiple tags, and COSI can assume less release breakage by
not relying on more advanced handling.

When the cloudbuild script detects a release is happening, it will now
apply only the semver tag to the image.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>