proto: update go modules
Update the go modules in proto. Dependabot is having trouble updating
proto dependencies because of use of protobuf/genproto's old URIs.
Update the URIs to use the new google.golang.org location, which
required changing the fake-gen import line for changed paths.

With these changes, all of proto's modules are able to be updated, and
sidecar (which relies on proto) was updated successfully as well.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>