ci: add prow lint/test scripts w/ CPU limits
Unit tests and golangci-lint are taking a long time to run in Prow
environments, sometimes timing out and causing CI flakes. Likely, this
is a result of the CI detecting the number of CPUs on the host, yet
being limited to fewer CPUs (2 currently) in the Prow job config.

Regardless of whether Prow jobs are updated to allow more CPU/memory
usage, it is still important to be able to limit concurrency to what is
configured in the job to avoid the issue if Prow machines become
significantly bigger in the future.

Prow jobs will need to be updated to make use of these new scripts.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>