prow: debug GOMAXPROCS instead of setting it
GOMAXPROCS is now being set in the prow job config. However, the issue
of unit tests and lint runs being throttled could still be a risk.
Instead of setting a hardcoded GOMAXPROCS value, log the value in the
test/lint runs so we can debug any potential issues in the future.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>