add .dockerignore file
Add a .dockerignore file that ignores hidden directories. On MacOS or
other environments that run a VM for container building, copying the
build context can take quite a bit of time.

Currently, ./.git, ./.cache, ./proto/.build, and ./proto/.protoc amount
to over 150MB of context that is not needed for builds to succed.

The vendor dir is still passed to the build context.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>