add kube-api-linter and resolve API issues
Add kube-api-linter to `make lint`.

kube-api-linter is based on golangci-lint v2, so it is upgraded to the
latest version.

kube-api-linter could be integrated with COSI CI in many ways. After
experimentation, it is fastest to run it as a standalone install rather
than build it each time `make lint` is run.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>